+++ src/oca_github_bot/tasks/merge_bot.py
@@ -13,7 +13,7 @@ from ..config import (
     SIMPLE_INDEX_ROOT,
     switchable,
 )
-from ..manifest import bump_manifest_version, git_modified_addon_dirs
+from ..manifest import bump_manifest_version, git_modified_addon_dirs, is_maintainer
 from ..queue import getLogger, task
 from ..version_branch import make_merge_bot_branch, parse_merge_bot_branch
 from .main_branch_bot import main_branch_bot_actions
@@ -79,7 +79,7 @@ def _merge_bot_merge_pr(org, repo, merge_bot_branch, dry_run=False):
     # to the PR.
     _git_call(["git", "fetch", "origin", f"refs/pull/{pr}/head:tmp-pr-{pr}"])
     _git_call(["git", "checkout", f"tmp-pr-{pr}"])
-    modified_addon_dirs = git_modified_addon_dirs(".", target_branch)
+    modified_addon_dirs, _ = git_modified_addon_dirs(".", target_branch)
     # Run main branch bot actions before bump version.
     # Do not run the main branch bot if there are no modified addons,
     # because it is dedicated to addons repos.
@@ -128,20 +128,30 @@ def _merge_bot_merge_pr(org, repo, merge_bot_branch, dry_run=False):
     return True
 
 
+def _user_can_merge(gh, org, repo, username, addons_dir, target_branch):
+    """
+    Check if a user is allowed to merge.
+
+    addons_dir must be a git clone of the branch to merge to target_branch.
+    """
+    gh_repo = gh.repository(org, repo)
+    if github.github_user_can_push(gh_repo, username):
+        return True
+    modified_addon_dirs, other_changes = git_modified_addon_dirs(
+        addons_dir, target_branch
+    )
+    if other_changes or not modified_addon_dirs:
+        return False
+    # if we are modifying addons only, then the user must be maintainer of all of them
+    return is_maintainer(username, modified_addon_dirs)
+
+
 @task()
 @switchable("merge_bot")
 def merge_bot_start(
     org, repo, pr, username, bumpversion=None, dry_run=False, intro_message=None
 ):
     with github.login() as gh:
-        if not github.github_user_can_push(gh.repository(org, repo), username):
-            gh_pr = gh.pull_request(org, repo, pr)
-            github.gh_call(
-                gh_pr.create_comment,
-                f"Sorry @{username} "
-                f"you do not have push permission, so I can't merge for you.",
-            )
-            return
         gh_pr = gh.pull_request(org, repo, pr)
         target_branch = gh_pr.base.ref
         try:
@@ -154,6 +164,21 @@ def merge_bot_start(
                     ["git", "fetch", "origin", f"pull/{pr}/head:{merge_bot_branch}"]
                 )
                 _git_call(["git", "checkout", merge_bot_branch])
+                if not _user_can_merge(gh, org, repo, username, ".", target_branch):
+                    github.gh_call(
+                        gh_pr.create_comment,
+                        f"Sorry @{username} you are not allowed to merge.\n\n"
+                        f"To do so you must either have push permissions on "
+                        f"the repository, or be a declared maintainer of all "
+                        f"modified addons.\n\n."
+                        f"If you wish to adopt an addon and become it's "
+                        f"[maintainer]"
+                        f"(https://odoo-community.org/page/maintainer-role), "
+                        f"open a pull request to add "
+                        f"your GitHub login to the `maintainers` key of its "
+                        f"manifest.",
+                    )
+                    return
                 # TODO for each modified addon, wlc lock / commit / push
                 # TODO then pull target_branch again
                 _git_call(["git", "rebase", "--autosquash", target_branch])